Fix SysGettid for 32 bit systems.
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@5942 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ incubator/nixclib/unixapi.cpp
@@ -409,7 +409,7 @@ RexxRoutine0(RexxObjectPtr,
              SysGettid)
 {
     pthread_t tid = pthread_self();
-    return context->UnsignedInt64ToObject(tid);
+    return context->UnsignedInt64ToObject((uint64_t)tid);
 }
